+++ lib/views/game/magi-mixer/page/index.tsx
@@ -1,0 +1,1 @@
+export * from "./page";
